Implemented a display format --stat for commits.
The format can be reached by using a/i in a commit object. <CR> on one of the
Files opens a Gdiff for the file and the current commit object.

I should probably extend this with a nowrap or provide some other method then
simply adding width=300 to make sure the complete filename can be expanded.